WellPredict/requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.9.0
networkx
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
prov==2.0.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.2
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.11
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
referencing==0.35.1
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.1
scipy
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sounddevice==0.5.0
SpeechRecognition==3.10.4
streamlit==1.39.0
streamlit-option-menu==0.3.13
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.44.0

plotly==5.22.0
